queries 패키지화 및 그에 따른 handler 수정
